+++ test/spec/configure.spec.ts
@@ -1,3 +1,4 @@
+import path = require('path');
 import { Configuration, configure, UndeclaredKeyError } from '../../src';
 
 class Environment extends Configuration {
@@ -6,6 +7,9 @@ class Environment extends Configuration {
 
 class ProcessEnvTestEnvironment extends Configuration {
 	readonly PROC_ENV = 'test';
+
+	readonly TEST_READ_PROCESS_ENV: string;
+	readonly TEST_DISABLE_READ_PROCESS_ENV: string;
 }
 
 class ProcessEnvTest2Environment extends Configuration {
@@ -27,9 +31,67 @@ describe('configure', () => {
 		})).toThrow(new UndeclaredKeyError(expectedErrorMessage));
 	});
 
-	it('rehydrates process.env', () => {
-		expect(process.env.PROC_ENV).toBe(undefined);
+	it('can accept an absolute path without a filename', () => {
+		const env: Environment = configure(Environment, {
+			absolutePath: path.join(process.cwd(), 'test', 'envs'),
+			overwriteProcessEnv: false,
+			fromProcessEnv: false,
+		});
 
+		expect(env.APP_TITLE).toBe('Default filename');
+	});
+
+	it('can accept an absolute path with a filename', () => {
+		const env: Environment = configure(Environment, {
+			absolutePath: path.join(
+				process.cwd(),
+				'test/envs/absolute.env'
+			),
+			overwriteProcessEnv: false,
+			fromProcessEnv: false,
+		});
+
+		expect(env.APP_TITLE).toBe('Absolute Path');
+	});
+
+	it('can accept a relative path without a filename', () => {
+		const env: Environment = configure(Environment, {
+			relativePath: 'test/envs/',
+			overwriteProcessEnv: false,
+			fromProcessEnv: false,
+		});
+
+		expect(env.APP_TITLE).toBe('Default filename');
+	});
+
+	it('can accept a relative path with a filename', () => {
+		const env: Environment = configure(Environment, {
+			relativePath: 'test/envs/relative.env',
+			overwriteProcessEnv: false,
+			fromProcessEnv: false,
+		});
+
+		expect(env.APP_TITLE).toBe('Relative Path');
+	});
+
+	it('reads from process.env', () => {
+		process.env.TEST_READ_PROCESS_ENV = 'read';
+
+		const env = configure(ProcessEnvTestEnvironment);
+		expect(env.TEST_READ_PROCESS_ENV).toBe('read');
+	});
+
+	it('doesn\'t read from process.env when disabled', () => {
+		process.env.TEST_DISABLE_READ_PROCESS_ENV = 'read';
+
+		const env = configure(
+			ProcessEnvTestEnvironment,
+			{ fromProcessEnv: false }
+		);
+		expect(env.TEST_DISABLE_READ_PROCESS_ENV).not.toBe('read');
+	});
+
+	it('rehydrates process.env', () => {
 		configure(ProcessEnvTestEnvironment);
 		expect(process.env.PROC_ENV).toBe('test');
 	});